Redux refactor of Translate page
