+++ app/src/Words/Translate.js
@@ -1,16 +1,86 @@
 import React from 'react';
 import {connect} from 'react-redux';
+import {addWord} from './actions';
 
 class Translate extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
+      words: props.words,
+      translateTextarea: '',
+      translatedWords: [],
       showNewWordForm: false,
       newWord: {
         main: '',
         secret: ''
       }
     };
+  }
+
+  componentWillReceiveProps(nextProps) {
+    this.setState({
+      words: nextProps.words
+    }, () => this.updateTranslatedWords(this.state.translateTextarea));
+  }
+
+  onTranslateTextareaChange(event) {
+    let mainText = event.target.value;
+    this.setState({
+      translateTextarea: mainText
+    });
+    this.updateTranslatedWords(mainText);
+  }
+
+  updateTranslatedWords(mainText) {
+    let translatedWords = [];
+    mainText = mainText.trim().split(" ");
+
+    // Loop through the english textarea text
+    for (let i = 0; i < mainText.length; i++) {
+      let isWordInDictionary = false;     // Whether the word is in the dictionary or not.
+      let specialChar = "";  // The special character that is at the end of the word.
+
+      // Check if the word ends in a special character
+      let specialCharsList = [',', '.', ';', ':', '?', '!'];
+      if (specialCharsList.indexOf(mainText[i].charAt(mainText[i].length - 1)) !== -1) {
+        specialChar = mainText[i].charAt(mainText[i].length - 1);
+        mainText[i] = mainText[i].substring(0, mainText[i].length - 1);
+      }
+
+      // Loop through the dictionary and check if the word has been defined yet.
+      for (let j = 0; j < this.state.words.length; j++) {
+        // if it is found, add it to the translatedWords array
+        if (mainText[i].toUpperCase() === this.state.words[j].mainWord.toUpperCase()) {
+          // The word to add to the translated words array
+          let newWord = this.state.words[j].secretWord;
+          newWord += specialChar + ' ';
+          translatedWords.push({
+            text: newWord,
+            hasClick: false
+          });
+          this.setState({
+            translatedWords: translatedWords
+          });
+          isWordInDictionary = true;
+          break;
+        }
+      }
+      if (!isWordInDictionary) {
+        let newWord = mainText[i]; // The word to add to the translated words array.
+        translatedWords.push({
+          text: newWord,
+          hasClick: true
+        });
+        translatedWords.push({
+          text: specialChar +  ' ',
+          hasClick: false
+        });
+
+        this.setState({
+          translatedWords: translatedWords
+        });
+      }
+    }
   }
 
   handleUntranslatedWordClick(word) {
@@ -28,8 +98,19 @@ class Translate extends React.Component {
   }
 
   handleClearTranslationClick() {
-    this.setState({showNewWordForm: false});
-    this.props.onClearTranslationClick();
+    this.setState({
+      showNewWordForm: false,
+      translateTextarea: '',
+      translatedWords: []
+    });
+  }
+
+  handleNewWordSecretChange(event) {
+    let newWord = {...this.state.newWord};
+    newWord.secret = event.target.value;
+    this.setState({
+      newWord: newWord
+    });
   }
 
   handleNewWordFormSDubmit(event) {
@@ -42,34 +123,26 @@ class Translate extends React.Component {
         secret: ''
       }
     });
-    this.props.onNewWordFormSubmit(word);
-  }
-
-  handleNewWordSecretChange(event) {
-    let newWord = {...this.state.newWord};
-    newWord.secret = event.target.value;
-    this.setState({
-      newWord: newWord
-    });
+    this.props.onNewWordFormSubmit(word.main, word.secret);
   }
 
   render() {
     let translatedText = '';
-    if (this.props.translatedWords) {
-      translatedText = this.props.translatedWords.map((word, index) => {
+    if (this.state.translatedWords) {
+      translatedText = this.state.translatedWords.map((word, index) => {
         return (
           <span
             key={index}
-            onClick={word.hasClick ? (() => this.handleUntranslatedWordClick(word.word)) : undefined}
+            onClick={word.hasClick ? (() => this.handleUntranslatedWordClick(word.text)) : undefined}
             className={word.hasClick ? 'highlight' : ''}>
-            {word.word}
+            {word.text}
           </span>
         );
       });
     }
     return (
       <div>
-        <textarea rows="5" placeholder="Enter your text to be translated..." value={this.props.translateText} onChange={this.props.onTextareaChange} autoFocus></textarea>
+        <textarea rows="5" placeholder="Enter your text to be translated..." value={this.state.translateTextarea} onChange={(e) => this.onTranslateTextareaChange(e)} autoFocus></textarea>
         <div>
           {translatedText}
         </div>
@@ -96,16 +169,13 @@ class Translate extends React.Component {
 
 const mapStateToProps = (state, ownProps) => {
   return {
-    translateText: state.translateTextArea,
-    translatedWords: state.translatedWords
+    words: state.words
   }
 }
 
 const mapDispatchToProps = (dispatch, ownProps) => {
-  return {      
-    onTextareaChange: () => console.log('Translate onTextareaChange'),
-    onClearTranslationClick: () => console.log('Translate onClearTranslationClick'),
-    onNewWordFormSubmit: () => console.log('Translate onNewWordFormSubmit')
+  return {
+    onNewWordFormSubmit: (mainWord, secretWord) => dispatch(addWord(mainWord, secretWord))
   }
 }
